backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
boto3==1.35.44
botocore==1.35.44
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
Django==5.1.2
django-cors-headers==4.5.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
